+++ src/Plugins/SmartStore.PayPal/changelog.md
@@ -3,6 +3,7 @@
 ##Paypal 2.2.0.4
 ###New Features
 * Option for API security protocol
+* Option to display express checkout button in mini shopping cart
 ###Bugfixes
 * "The request was aborted: Could not create SSL/TLS secure channel." See https://devblog.paypal.com/upcoming-security-changes-notice/